create a collection of additional dict toolz
